+++ src/encode/CP932.test.js
@@ -21,21 +21,20 @@ const equalsArray = function(x, y) {
 	test("getWidthForCP932", () => { expect(CP932.getWidthForCP932("ABCあいう高髙①")).toBe(15); });
 }
 
-
 {
 	const target_text = "ABCあいう高髙①";
 	const cp932array = CP932.toCP932Array(target_text);
 	const cp932array_ = [0x0041, 0x0042, 0x0043, 0x82A0, 0x82A2, 0x82A4, 0x8D82, 0xFBFC, 0x8740];
 	test("toCP932Array", () => { expect(equalsArray(cp932array, cp932array_)).toBe(true); });
-	test("fromCP932Array", () => { expect(CP932.fromCP932Array(cp932array_)).toBe(target_text); });
+	test("fromCP932Array", () => { expect(CP932.fromCP932Array(cp932array_).encode_string).toBe(target_text); });
 }
 
 {
 	const target_text = "ABCあいう高髙①";
-	const cp932arraybin = CP932.toCP932ArrayBinary(target_text);
+	const cp932arraybin = CP932.toCP932Binary(target_text);
 	const cp932arraybin_ = [0x41, 0x42, 0x43, 0x82, 0xA0, 0x82, 0xA2, 0x82, 0xA4, 0x8D, 0x82, 0xFB, 0xFC, 0x87, 0x40];
 	test("toCP932ArrayBinary", () => { expect(equalsArray(cp932arraybin, cp932arraybin_)).toBe(true); });
-	test("fromCP932Array (binary)", () => { expect(CP932.fromCP932Array(cp932arraybin_)).toBe(target_text); });
+	test("fromCP932Array (binary)", () => { expect(CP932.fromCP932Array(cp932arraybin_).encode_string).toBe(target_text); });
 }
 
 {